Introduce statistics and fix combo logic
The introduction of statistics calculation is critical to detect any
discrepancy between the target PAR sheet and the implementation.

In this case, it was revealed that the actual "XX" symbols on the sheet
did *not* meant just blank symbols as implemented at the start, but
rather *any* other symbol than Jackpot.

Thus the payout has been readjusted and now matches the sheets specs:
https://easy.vegas/games/slots/par_sheets/generic-1987.gif